requirements.txt
appdirs==1.4.3
bidict==0.21.2
CacheControl==0.12.6
cachetools==4.2.1
certifi==2019.11.28
cffi==1.14.5
chardet==3.0.4
click==7.1.2
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
dnspython==1.16.0
eventlet==0.30.1
Flask==1.1.2
Flask-Cors==3.0.10
Flask-Login==0.5.0
Flask-SocketIO==5.0.1
google-api-core==1.26.0
google-auth==1.27.0
google-cloud-core==1.6.0
google-cloud-storage==1.36.0
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.52.0
greenlet==1.0.0
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
itsdangerous==1.1.0
Jinja2==2.11.3
lockfile==0.12.2
MarkupSafe==1.1.1
msgpack==0.6.2
packaging==20.3
pep517==0.8.2
progress==1.5
protobuf==3.15.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.6
python-dotenv==0.15.0
python-engineio==4.0.0
python-socketio==5.0.4
pytoml==0.1.21
pytz==2021.1
requests==2.22.0
retrying==1.3.3
rsa==4.7.1
six==1.14.0
urllib3==1.25.8
webencodings==0.5.1
Werkzeug==1.0.1
